rename to GameScreen and remove debug ToDo
